KTOR-9645 Fixup docs; extract duplication for engine selection
